Rename classes and files
