+++ test/test_control_server.py
@@ -27,7 +27,7 @@ try:
 except ImportError:
     from mock import Mock, patch, call
 
-from core.rpc_server import ExposedObject, ExposedObjectCollection, ZMQJSONRPCServer
+from core.control_server import ExposedObject, ExposedObjectCollection, ControlServer
 from zmq import Again as zmq_again_exception
 from zmq import NOBLOCK as zmq_no_block_flag
 
@@ -153,21 +153,21 @@ class TestExposedObjectCollection(unittest.TestCase):
         obj.getTest.assert_called_once_with(454, 43)
 
 
-class TestZMQJSONRPCServer(unittest.TestCase):
-    @patch('core.rpc_server.ZMQJSONRPCServer._get_zmq_rep_socket')
+class TestControlServer(unittest.TestCase):
+    @patch('core.control_server.ControlServer._get_zmq_rep_socket')
     def test_connection(self, mock_socket_method):
-        ZMQJSONRPCServer(host='127.0.0.1', port='10001')
+        ControlServer(host='127.0.0.1', port='10001')
 
         mock_socket_method.assert_has_calls([call(), call().bind('tcp://127.0.0.1:10001')])
 
-    @patch('core.rpc_server.ZMQJSONRPCServer._get_zmq_rep_socket')
+    @patch('core.control_server.ControlServer._get_zmq_rep_socket')
     def test_process_does_not_block(self, mock_socket_method):
         mock_socket = Mock()
         mock_socket.recv_unicode.side_effect = zmq_again_exception()
 
         mock_socket_method.return_value = mock_socket
 
-        server = ZMQJSONRPCServer()
+        server = ControlServer()
         assertRaisesNothing(self, server.process)
 
         mock_socket.recv_unicode.assert_has_calls([call(flags=zmq_no_block_flag)])
